+++ sdk/src/main/java/ly/count/android/sdk/ModuleContent.java
@@ -329,7 +329,8 @@ public class ModuleContent extends ModuleBase {
             orientation,
             globalContentCallback,
             this::notifyAfterContentIsClosed,
-            _cly.config_.content.allowedIntentSchemes
+            _cly.config_.content.allowedIntentSchemes,
+            _cly.config_.content.contentUrlHandler
         );
 
         contentOverlay.attachToActivity(activity);
@@ -404,6 +405,9 @@ public class ModuleContent extends ModuleBase {
         assert response != null;
 
         String content = response.optString("html");
+        if (!content.isEmpty()) {
+            content = UtilsDevice.appendThemeParam(content, _cly.context_);
+        }
         JSONObject coordinates = response.optJSONObject("geo");
 
         assert coordinates != null;